pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-info"
version = "0.1.0"
authors = [
    {name = "Ryan Manzer", email = "ryan.manzer@gmail.com"}
]
description = "A command line tool to extract exif metadata information from a photo."
readme = "README.md"
keywords = ["photography", "exif", "metadata", "cli", "markdown"]
license = {text = "GPL-3.0-or-later"}
dependencies = [
    "exif",
    "pillow",
    "typer",
    "rich",
    "tomli>=2.0.0",
]
requires-python = ">=3.8"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/supermanzer/photo-info"
Documentation = "https://github.com/supermanzer/photo-info#readme"
Repository = "https://github.com/supermanzer/photo-info.git"
Issues = "https://github.com/supermanzer/photo-info/issues"

[project.scripts]
photo-info = "photo_info.cli:app"

[tool.setuptools]
packages = ["photo_info"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
photo_info = ["py.typed"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]